+++ src/Application/Query/Rooming/RoomingList/Export/UserViewQueryHandler.php
@@ -13,6 +13,7 @@ namespace Proximum\Vimeet\Application\Query\Rooming\RoomingList\Export;
 use Proximum\Vimeet\Application\View\Rooming\ExportList\UserSheetView;
 use Proximum\Vimeet\Domain\Model\Spot;
 use Proximum\Vimeet\Domain\Model\User\Event\ExtraData;
+use Proximum\Vimeet\Domain\Order\Merger;
 use Proximum\Vimeet\Domain\Repository\SheetRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\User\Event\ExtraDataRepositoryInterface;
 use Proximum\Vimeet\Domain\User\Event\ExtraData\Type;
@@ -25,12 +26,17 @@ class UserViewQueryHandler
     /** @var ExtraDataRepositoryInterface */
     private $extraDataRepository;
 
+    /** @var Merger */
+    private $merger;
+
     public function __construct(
         SheetRepositoryInterface $sheetRepository,
-        ExtraDataRepositoryInterface $extraDataRepository
+        ExtraDataRepositoryInterface $extraDataRepository,
+        Merger $merger
     ) {
         $this->sheetRepository = $sheetRepository;
         $this->extraDataRepository = $extraDataRepository;
+        $this->merger = $merger;
     }
 
     public function handle(UserViewQuery $query): UserSheetView
@@ -59,6 +65,8 @@ class UserViewQueryHandler
 
         $sheetIds = [];
         $sheetTitles = [];
+        $sheetFollowers = [];
+        $sheetPlans = [];
         $typeTitles = [];
         $spotReferences = [];
 
@@ -70,6 +78,18 @@ class UserViewQueryHandler
         foreach ($sheets as $sheet) {
             $sheetIds[] = $sheet->getId();
             $sheetTitles[] = $sheet->getTitle();
+            $sheetFollowers[] = $sheet->getFollowerName();
+
+            $mergedOrder = null;
+            $plan = null;
+
+            if (null !== $this->merger->getMergedOrders($sheet)) {
+                $mergedOrder = $this->merger->getMergedOrders($sheet);
+                $plan = $mergedOrder->getPlan();
+                if (null !== $plan) {
+                    $sheetPlans[] = $plan->getName();
+                }
+            }
 
             if (!isset($typeTitles[$sheet->getType()->getId()])) {
                 $typeTitles[$sheet->getType()->getId()] = $sheet->getTypeTitle($query->locale);
@@ -82,11 +102,15 @@ class UserViewQueryHandler
 
         $userSheetView = new UserSheetView(
             $query->user->getId(),
-            $query->user->getAccount()->getGender(),
-            $query->user->getAccount()->getFirstName(),
-            $query->user->getAccount()->getLastName(),
+            $query->user->getAccount()->getGender() ?? '',
+            $query->user->getAccount()->getFirstName() ?? '',
+            $query->user->getAccount()->getLastName() ?? '',
+            $query->user->getEmail(),
+            $query->user->getMobile() ?? '',
             implode(',', $sheetIds),
             implode(',', $sheetTitles),
+            implode(',', $sheetFollowers),
+            implode(',', $sheetPlans),
             implode(',', $typeTitles),
             implode(',', $spotReferences),
             $comment,