+++ tests/Application/Query/Rooming/RoomingList/Export/UserViewQueryHandlerTest.php
@@ -15,10 +15,13 @@ use Proximum\Vimeet\Application\Query\Rooming\RoomingList\Export\UserViewQueryHa
 use PHPUnit\Framework\TestCase;
 use Proximum\Vimeet\Application\View\Rooming\ExportList\UserSheetView;
 use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Domain\Model\Order;
+use Proximum\Vimeet\Domain\Model\Product;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\Spot;
 use Proximum\Vimeet\Domain\Model\Type;
 use Proximum\Vimeet\Domain\Model\User;
+use Proximum\Vimeet\Domain\Order\Merger;
 use Proximum\Vimeet\Domain\Repository\SheetRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\User\Event\ExtraDataRepositoryInterface;
 use Proximum\Vimeet\Domain\User\Event\ExtraData\Type as ExtraDataType;
@@ -37,6 +40,24 @@ class UserViewQueryHandlerTest extends TestCase
         $spot = $this->prophesize(Spot::class);
         $type1 = $this->prophesize(Type::class);
         $type2 = $this->prophesize(Type::class);
+
+        $merger = $this->prophesize(Merger::class);
+
+        $order1 = $this->prophesize(Order::class);
+        $order2 = $this->prophesize(Order::class);
+        $order3 = $this->prophesize(Order::class);
+
+        $plan1 = $this->prophesize(Product::class);
+        $plan2 = $this->prophesize(Product::class);
+        $plan3 = $this->prophesize(Product::class);
+
+        $merger->getMergedOrders($sheet1->reveal())->shouldBeCalled()->willReturn($order1->reveal());
+        $merger->getMergedOrders($sheet2->reveal())->shouldBeCalled()->willReturn($order2->reveal());
+        $merger->getMergedOrders($sheet3->reveal())->shouldBeCalled()->willReturn($order3->reveal());
+
+        $order1->getPlan()->shouldBeCalled()->willReturn($plan1->reveal());
+        $order2->getPlan()->shouldBeCalled()->willReturn($plan2->reveal());
+        $order3->getPlan()->shouldBeCalled()->willReturn($plan3->reveal());
 
         $commentExtraData = $this->prophesize(User\Event\ExtraData::class);
         $tastingExtraData = $this->prophesize(User\Event\ExtraData::class);
@@ -66,9 +87,19 @@ class UserViewQueryHandlerTest extends TestCase
         $sheet2->getSpot()->shouldBeCalled()->willReturn(null);
         $sheet3->getSpot()->shouldBeCalled()->willReturn($spot->reveal());
 
+        $sheet1->getFollowerName()->shouldBeCalled()->willReturn('Al Pacino');
+        $sheet2->getFollowerName()->shouldBeCalled()->willReturn('Robert DeNiro');
+        $sheet3->getFollowerName()->shouldBeCalled()->willReturn('Joe Pesci');
+
+        $plan1->getName()->shouldBeCalled()->willReturn('Plan 1');
+        $plan2->getName()->shouldBeCalled()->willReturn('Plan 2');
+        $plan3->getName()->shouldBeCalled()->willReturn('Plan 3');
+
         $spot->getReference()->shouldBeCalled()->willReturn('A123');
 
         $user->getId()->shouldBeCalled()->willReturn(1);
+        $user->getEmail()->shouldBeCalled()->willReturn('test@test.com');
+        $user->getMobile()->shouldBeCalled()->willReturn('0000000001');
         $user->getAccount()->shouldBeCalled()->willReturn($account->reveal());
 
         $account->getGender()->shouldBeCalled()->willReturn('man');
@@ -98,7 +129,8 @@ class UserViewQueryHandlerTest extends TestCase
         $query = new UserViewQuery($event->reveal(), $user->reveal(), $locale);
         $handler = new UserViewQueryHandler(
             $sheetRepository->reveal(),
-            $extraDataRepository->reveal()
+            $extraDataRepository->reveal(),
+            $merger->reveal()
         );
 
         $result = $handler->handle($query);
@@ -108,8 +140,12 @@ class UserViewQueryHandlerTest extends TestCase
             'man',
             'Jean',
             'Paul',
+            'test@test.com',
+            '0000000001',
             '1,2,3',
             'Aanera,Bbnera,Ccnera',
+            'Al Pacino,Robert DeNiro,Joe Pesci',
+            'Plan 1,Plan 2,Plan 3',
             'Exposant,Visiteur',
             'A123',
             'This is a comment',