+++ tests/Infrastructure/Bundle/InfrastructureBundle/Serializer/Normalizer/Rooming/ExportList/RoomingListViewNormalizerTest.php
@@ -29,8 +29,12 @@ class RoomingListViewNormalizerTest extends TestCase
             'man',
             'Jean',
             'Paul',
+            'jean.paul@test.com',
+            '0101010101',
             '1,2,3',
             'Aanera,Bbnera,Ccnera',
+            '',
+            'Plan 1,Plan 2,Plan 3',
             'Exposant,Visiteur',
             'A123',
             'This is a comment',
@@ -41,8 +45,12 @@ class RoomingListViewNormalizerTest extends TestCase
             'woman',
             'Marie',
             'Curie',
+            'marie.curie@test.com',
+            '0202020202',
             '1',
             'Aanera',
+            'Alfred Einstein,Anatol Tesla',
+            'Plan 3',
             'Exposant',
             '',
             'No comment',
@@ -53,8 +61,12 @@ class RoomingListViewNormalizerTest extends TestCase
             'man',
             'Jean',
             'Paul',
+            'paul.jean@test.com',
+            '0303030303',
             '4,5',
             'Lorem,Ipsum',
+            '',
+            'Plan 1,Plan 2',
             'Exposant',
             'A321',
             '',
@@ -65,8 +77,12 @@ class RoomingListViewNormalizerTest extends TestCase
             '',
             'Bidule',
             'Truc',
+            'bidule.truc@test.com',
+            '0404040404',
             '5',
             'Aanera,Bbnera,Ccnera',
+            'Jean Claude VanDamme',
+            'Plan 2',
             'Exposant,Visiteur',
             '',
             'A comment',
@@ -100,6 +116,10 @@ class RoomingListViewNormalizerTest extends TestCase
             ->shouldBeCalled()
             ->willReturn('typeTitle');
         $translator
+            ->trans('rooming_list_data_export.column.spotReference', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('spotReference');
+        $translator
             ->trans('rooming_list_data_export.column.userId', [], 'export', 'fr')
             ->shouldBeCalled()
             ->willReturn('userId');
@@ -116,10 +136,6 @@ class RoomingListViewNormalizerTest extends TestCase
             ->shouldBeCalled()
             ->willReturn('userLastName');
 
-        $translator
-            ->trans('rooming_list_data_export.column.spotReference', [], 'export', 'fr')
-            ->shouldBeCalled()
-            ->willReturn('spotReference');
         $translator
             ->trans('rooming_list_data_export.column.userComment', [], 'export', 'fr')
             ->shouldBeCalled()
@@ -140,9 +156,19 @@ class RoomingListViewNormalizerTest extends TestCase
             ->willReturn('roomType')
         ;
         $translator
+            ->trans('rooming_list_data_export.column.sheetFollower', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('sheetFollower')
+        ;
+        $translator
             ->trans('rooming_list_data_export.column.roomNumber', [], 'export', 'fr')
             ->shouldBeCalled()
             ->willReturn('roomNumber')
+        ;
+        $translator
+            ->trans('rooming_list_data_export.column.roomPlan', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('roomPlan')
         ;
          $translator
             ->trans('rooming_list_data_export.column.arrival', [], 'export', 'fr')
@@ -153,6 +179,26 @@ class RoomingListViewNormalizerTest extends TestCase
             ->trans('rooming_list_data_export.column.departure', [], 'export', 'fr')
             ->shouldBeCalled()
             ->willReturn('departure')
+        ;
+        $translator
+            ->trans('rooming_list_data_export.column.email', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('email')
+        ;
+        $translator
+            ->trans('rooming_list_data_export.column.mobile', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('mobile')
+        ;
+        $translator
+            ->trans('rooming_list_data_export.column.roommate.roomPlan', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('roommate.roomPlan')
+        ;
+        $translator
+            ->trans('rooming_list_data_export.column.roommate.sheetFollower', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('roommate.sheetFollower')
         ;
 
         $translator
@@ -183,6 +229,14 @@ class RoomingListViewNormalizerTest extends TestCase
             ->trans('rooming_list_data_export.column.roommate.userLastName', [], 'export', 'fr')
             ->shouldBeCalled()
             ->willReturn('roommate.userLastName');
+        $translator
+            ->trans('rooming_list_data_export.column.roommate.mobile', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('roommate.mobile');
+        $translator
+            ->trans('rooming_list_data_export.column.roommate.email', [], 'export', 'fr')
+            ->shouldBeCalled()
+            ->willReturn('roommate.email');
         $translator
             ->trans('rooming_list_data_export.column.roommate.spotReference', [], 'export', 'fr')
             ->shouldBeCalled()
@@ -236,12 +290,12 @@ class RoomingListViewNormalizerTest extends TestCase
 
         $result = $serializer->serialize($this->getInput(), 'csv', ['csv_delimiter' => ';']);
 
-        $expected = 'sheetId;sheetTitle;typeTitle;spotReference;userId;userGender;userFirstName;userLastName;userComment;userTasting;accommodationTitle;roomType;arrival;departure;roomNumber;roommate.sheetId;roommate.sheetTitle;roommate.typeTitle;roommate.spotReference;roommate.userId;roommate.userGender;roommate.userFirstName;roommate.userLastName;roommate.userComment;roommate.userTasting
-rooming_list_data_export.column.sheetId;rooming_list_data_export.column.sheetTitle;rooming_list_data_export.column.typeTitle;rooming_list_data_export.column.spotReference;rooming_list_data_export.column.userId;rooming_list_data_export.column.userGender;rooming_list_data_export.column.userFirstName;rooming_list_data_export.column.userLastName;rooming_list_data_export.column.userComment;rooming_list_data_export.column.userTasting;rooming_list_data_export.column.accommodationTitle;rooming_list_data_export.column.roomType;rooming_list_data_export.column.arrival;rooming_list_data_export.column.departure;rooming_list_data_export.column.roomNumber;rooming_list_data_export.column.roommate.sheetId;rooming_list_data_export.column.roommate.sheetTitle;rooming_list_data_export.column.roommate.typeTitle;rooming_list_data_export.column.roommate.spotReference;rooming_list_data_export.column.roommate.userId;rooming_list_data_export.column.roommate.userGender;rooming_list_data_export.column.roommate.userFirstName;rooming_list_data_export.column.roommate.userLastName;rooming_list_data_export.column.roommate.userComment;rooming_list_data_export.column.roommate.userTasting
-1,2,3;Aanera,Bbnera,Ccnera;Exposant,Visiteur;A123;1;gender.man;Jean;Paul;"This is a comment";"This is a tasting";Mariott;rooming_list_data_export.column.roomType.single;08/01/2019;10/01/2019;A123;;;;;;;;;;
-1;Aanera;Exposant;;2;gender.woman;Marie;Curie;"No comment";;Mariott;rooming_list_data_export.column.roomType.single;08/01/2019;12/01/2019;A124;;;;;;;;;;
-4,5;Lorem,Ipsum;Exposant;A321;3;gender.man;Jean;Paul;;;Mariott;rooming_list_data_export.column.roomType.double;08/01/2019;12/01/2019;A125;5;Aanera,Bbnera,Ccnera;Exposant,Visiteur;;4;;Bidule;Truc;"A comment";"A tasting info"
-1,2,3;Aanera,Bbnera,Ccnera;Exposant,Visiteur;A123;1;gender.man;Jean;Paul;"This is a comment";"This is a tasting";Novotel;rooming_list_data_export.column.roomType.single;10/01/2019;12/01/2019;A126;;;;;;;;;;
+        $expected = 'sheetId;sheetTitle;sheetFollower;roomPlan;typeTitle;spotReference;userId;userGender;userFirstName;userLastName;email;mobile;userComment;userTasting;accommodationTitle;roomType;arrival;departure;roomNumber;roommate.sheetId;roommate.sheetTitle;roommate.sheetFollower;roommate.roomPlan;roommate.typeTitle;roommate.spotReference;roommate.userId;roommate.userGender;roommate.userFirstName;roommate.userLastName;roommate.email;roommate.mobile;roommate.userComment;roommate.userTasting
+rooming_list_data_export.column.sheetId;rooming_list_data_export.column.sheetTitle;rooming_list_data_export.column.sheetFollower;rooming_list_data_export.column.roomPlan;rooming_list_data_export.column.typeTitle;rooming_list_data_export.column.spotReference;rooming_list_data_export.column.userId;rooming_list_data_export.column.userGender;rooming_list_data_export.column.userFirstName;rooming_list_data_export.column.userLastName;rooming_list_data_export.column.email;rooming_list_data_export.column.mobile;rooming_list_data_export.column.userComment;rooming_list_data_export.column.userTasting;rooming_list_data_export.column.accommodationTitle;rooming_list_data_export.column.roomType;rooming_list_data_export.column.arrival;rooming_list_data_export.column.departure;rooming_list_data_export.column.roomNumber;rooming_list_data_export.column.roommate.sheetId;rooming_list_data_export.column.roommate.sheetTitle;rooming_list_data_export.column.roommate.sheetFollower;rooming_list_data_export.column.roommate.roomPlan;rooming_list_data_export.column.roommate.typeTitle;rooming_list_data_export.column.roommate.spotReference;rooming_list_data_export.column.roommate.userId;rooming_list_data_export.column.roommate.userGender;rooming_list_data_export.column.roommate.userFirstName;rooming_list_data_export.column.roommate.userLastName;rooming_list_data_export.column.roommate.email;rooming_list_data_export.column.roommate.mobile;rooming_list_data_export.column.roommate.userComment;rooming_list_data_export.column.roommate.userTasting
+1,2,3;Aanera,Bbnera,Ccnera;;"Plan 1,Plan 2,Plan 3";Exposant,Visiteur;A123;1;gender.man;Jean;Paul;jean.paul@test.com;0101010101;"This is a comment";"This is a tasting";Mariott;rooming_list_data_export.column.roomType.single;08/01/2019;10/01/2019;A123;;;;;;;;;;;;;;
+1;Aanera;"Alfred Einstein,Anatol Tesla";"Plan 3";Exposant;;2;gender.woman;Marie;Curie;marie.curie@test.com;0202020202;"No comment";;Mariott;rooming_list_data_export.column.roomType.single;08/01/2019;12/01/2019;A124;;;;;;;;;;;;;;
+4,5;Lorem,Ipsum;;"Plan 1,Plan 2";Exposant;A321;3;gender.man;Jean;Paul;paul.jean@test.com;0303030303;;;Mariott;rooming_list_data_export.column.roomType.double;08/01/2019;12/01/2019;A125;5;Aanera,Bbnera,Ccnera;"Jean Claude VanDamme";"Plan 2";Exposant,Visiteur;;4;;Bidule;Truc;bidule.truc@test.com;0404040404;"A comment";"A tasting info"
+1,2,3;Aanera,Bbnera,Ccnera;;"Plan 1,Plan 2,Plan 3";Exposant,Visiteur;A123;1;gender.man;Jean;Paul;jean.paul@test.com;0101010101;"This is a comment";"This is a tasting";Novotel;rooming_list_data_export.column.roomType.single;10/01/2019;12/01/2019;A126;;;;;;;;;;;;;;
 ';
 
         $this->assertEquals($expected, $result);